Changed the labels in MainPage so that when they're tapped, they use the SingleClick class. Also, added a question in CharacterDeletionPage to make sure the user wants to delete the character they select.

+++ StarWRPG/StarWRPG/Views/CharacterSelection/CharacterDeletionPage.cs
@@ -21,8 +21,12 @@ namespace StarWRPG.Views
         {
             if (e.SelectedItem is FFGCharacterViewModel character)
             {
-                await App.CharacterDatabase.DeleteCharacterAsync(character);
-                await Navigation.PopAsync();
+                var answer = await DisplayAlert("Delete Character", $"Are you sure you want to delete {character.Name}?", "Yes", "No");
+                if (answer)
+                {
+                    await App.CharacterDatabase.DeleteCharacterAsync(character);
+                    await Navigation.PopAsync();
+                }
             }
         }
     }
